feat: rename Entity\UserOwned trait and interface
