+++ src/Entity/Project/Project.php
@@ -7,15 +7,15 @@ use App\Entity\Category;
 use App\Entity\DateCreatedTrait;
 use App\Entity\DateUpdatedTrait;
 use App\Entity\Interface\AccountingOwnerInterface;
-use App\Entity\Interface\UserOwnedInterface;
 use App\Entity\LocalizedInterface;
 use App\Entity\LocalizedTrait;
 use App\Entity\Matchfunding\MatchCallSubmission;
 use App\Entity\Matchfunding\MatchCallSubmissionStatus;
 use App\Entity\MigratedTrait;
 use App\Entity\Territory;
-use App\Entity\Trait\UserOwnedTrait;
 use App\Entity\User\User;
+use App\Entity\UserOwnedInterface;
+use App\Entity\UserOwnedTrait;
 use App\Mapping\Provider\EntityMapProvider;
 use App\Repository\Project\ProjectRepository;
 use AutoMapper\Attribute\MapProvider;
