+++ src/Entity/Project/RewardClaim.php
@@ -3,9 +3,9 @@
 namespace App\Entity\Project;
 
 use App\Entity\Gateway\Charge;
-use App\Entity\Interface\UserOwnedInterface;
-use App\Entity\Trait\UserOwnedTrait;
 use App\Entity\User\User;
+use App\Entity\UserOwnedInterface;
+use App\Entity\UserOwnedTrait;
 use App\Mapping\Provider\EntityMapProvider;
 use App\Repository\Project\RewardClaimRepository;
 use AutoMapper\Attribute\MapProvider;
